플레이어 위치 패킷전송을 일정주기마다 보내주도록 변경
클라이언트로부터 플레이어의 위치정보를 받아 적용만하고 다시 다른 플레이어들에게 보내주는 것을 일정주기마다 하도록 수정하였다.
이렇게 수정함으로써 한번에 여러 플레이어의 위치 정보를 묶어서 보낼 수 있기 때문에 패킷 전송횟수를 줄일 수 있다.

+++ AvoidHimServer/Timer.h
@@ -19,7 +19,7 @@ public:
 	Timer();
 	~Timer();
 
-	double GetTimeElapsed() const;
+	float GetTimeElapsed() const;
 	int GetFPS() const;
 	bool IsPaused() const;
